requirements.txt
fastapi
uvicorn
httpx
pydantic-settings
pytest
pytest-asyncio
pytest-cov
sqlalchemy[asyncio]
asyncpg
alembic
pytz